Moved Lelantus from ecc_test into lelantus.h/cpp
