+++ core/CMakeLists.txt
@@ -13,6 +13,7 @@ set(CORE_SRC
     merkle.cpp
     negotiator.cpp
     lightning.cpp
+    lelantus.cpp
     proto.cpp
     peer_manager.cpp
     fly_client.cpp
